+++ src/main/java/kr/codesqaud/cafe/controller/ExceptionController.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.controller;
 
-import kr.codesqaud.cafe.AddressException;
+import kr.codesqaud.cafe.exception.AddressException;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
