+++ src/main/java/kr/codesqaud/cafe/exception/AddressException.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe;
+package kr.codesqaud.cafe.exception;
 
 public class AddressException extends RuntimeException {
     public AddressException(String message) {
